+++ src/main/java/net/geforcemods/securitycraft/blockentities/RetinalScannerBlockEntity.java
@@ -14,6 +14,7 @@ import net.geforcemods.securitycraft.api.Option.BooleanOption;
 import net.geforcemods.securitycraft.api.Option.DisabledOption;
 import net.geforcemods.securitycraft.api.Option.DoubleOption;
 import net.geforcemods.securitycraft.api.Option.IntOption;
+import net.geforcemods.securitycraft.api.Option.RespectInvisibilityOption;
 import net.geforcemods.securitycraft.api.Option.SignalLengthOption;
 import net.geforcemods.securitycraft.api.Owner;
 import net.geforcemods.securitycraft.blocks.RetinalScannerBlock;
@@ -50,6 +51,7 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 		}
 	};
 	private DisabledOption disabled = new DisabledOption(false);
+	private RespectInvisibilityOption respectInvisibility = new RespectInvisibilityOption();
 	private ResolvableProfile ownerProfile;
 	private int viewCooldown = 0;
 
@@ -72,7 +74,7 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 
 			int signalLength = getSignalLength();
 
-			if ((!state.getValue(RetinalScannerBlock.POWERED) || signalLength == 0) && !Utils.isEntityInvisible(entity)) {
+			if ((!state.getValue(RetinalScannerBlock.POWERED) || signalLength == 0) && !isConsideredInvisible(entity)) {
 				if (entity instanceof Player player) {
 					Owner viewingPlayer = new Owner(player);
 
@@ -168,7 +170,7 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 	@Override
 	public Option<?>[] customOptions() {
 		return new Option[] {
-				activatedByEntities, sendMessage, signalLength, disabled, maximumDistance
+				activatedByEntities, sendMessage, signalLength, disabled, maximumDistance, respectInvisibility
 		};
 	}
 
@@ -219,5 +221,10 @@ public class RetinalScannerBlockEntity extends DisguisableBlockEntity implements
 
 	public ResolvableProfile getPlayerProfile() {
 		return ownerProfile;
+	}
+
+	@Override
+	public boolean isConsideredInvisible(LivingEntity entity) {
+		return respectInvisibility.isConsideredInvisible(entity);
 	}
 }